feat: Add analysis components including Skeleton, Overlay, LoadingOverlay, ChartCard, Card, and FilterControls
