+++ src/components/Analysis/LoadingOverlay.jsx
@@ -1,0 +1,14 @@
+const LoadingOverlay = ({ loading }) => {
+    if (!loading) return null
+  
+    return (
+      <div className="fixed inset-0 bg-white/80 backdrop-blur-sm z-50 flex items-center justify-center">
+        <div className="animate-pulse flex flex-col items-center">
+          <div className="w-16 h-16 bg-primary/20 rounded-full mb-4"></div>
+          <p className="text-gray-600">Loading dashboard data...</p>
+        </div>
+      </div>
+    )
+  }
+  
+  export default LoadingOverlay
