+++ src/components/Analysis/Skeleton.jsx
@@ -1,0 +1,5 @@
+const Skeleton = ({ className = "" }) => {
+    return <div className={`animate-pulse bg-gray-200 rounded ${className}`}></div>
+  }
+  
+  export default Skeleton
