+++ src/pages/AnalysisPage.jsx
@@ -1,4 +1,4 @@
-import { useState, useEffect, use } from "react"
+import { useState, useEffect } from "react"
 import {
   FiUsers,
   FiBriefcase,
@@ -6,8 +6,6 @@ import {
   FiTrendingDown,
   FiBarChart2,
   FiPieChart,
-  FiMaximize2,
-  FiMinimize2,
   FiCalendar,
 } from "react-icons/fi"
 import {
@@ -34,6 +32,12 @@ import SourceOfApplication from "../components/AnalysisComponents/SourceOfApplic
 import HiringFunnelChart from "../components/AnalysisComponents/HiringFunnelChart"
 import BlacklistedRejected from "../components/AnalysisComponents/BlacklistedRejected"
 
+import FilterControls from "../components/Analysis/FilterControls"
+import Card from "../components/Analysis/Card"
+import ChartCard from "../components/Analysis/ChartCard"
+import Overlay from "../components/Analysis/Overlay"
+import Skeleton from "../components/Analysis/Skeleton"
+
 import api from "../api/axios"
 
 // Register ChartJS components
@@ -59,39 +63,13 @@ const getMonthName = (monthNum) => {
 // Custom colors for charts - teal palette
 const COLORS = ["#004040", "#006060", "#008080", "#00a0a0", "#00bfbf", "#e0f7fa", "#e6ffff"]
 
-// Custom Skeleton component
-const Skeleton = ({ className = "" }) => {
-  return <div className={`animate-pulse bg-gray-200 rounded ${className}`}></div>
-}
-
 const AnalysisPage = () => {
   const [expandedCard, setExpandedCard] = useState(null)
   const [loading, setLoading] = useState(true)
   const [year, setYear] = useState(new Date().getFullYear().toString())
-  const [month, setMonth] = useState("") // Add this line for month state
-
-  // Add this array for month options
-  const monthOptions = [
-    { value: "", label: "All Months" },
-    { value: "1", label: "January" },
-    { value: "2", label: "February" },
-    { value: "3", label: "March" },
-    { value: "4", label: "April" },
-    { value: "5", label: "May" },
-    { value: "6", label: "June" },
-    { value: "7", label: "July" },
-    { value: "8", label: "August" },
-    { value: "9", label: "September" },
-    { value: "10", label: "October" },
-    { value: "11", label: "November" },
-    { value: "12", label: "December" },
-  ]
-
-  // Add this array for year options (current year and 4 previous years)
-  const yearOptions = Array.from({ length: 5 }, (_, i) => {
-    const yearValue = (new Date().getFullYear() - i).toString()
-    return { value: yearValue, label: yearValue }
-  })
+  const [month, setMonth] = useState("")
+  const [selectedPosition, setSelectedPosition] = useState("")
+  const [positionsFilter, setPositionsFilter] = useState(null)
 
   // State for all dashboard data
   const [summaryData, setSummaryData] = useState(null)
@@ -100,28 +78,30 @@ const AnalysisPage = () => {
   const [applicationSource, setApplicationSource] = useState([])
   const [monthlyTrends, setMonthlyTrends] = useState([])
   const [jobPositions, setJobPositions] = useState([])
-  const [positionsFilter, setPositionsFilter] = useState(null)
   const [hiringFunnel, setHiringFunnel] = useState(null)
   const [recentApplicants, setRecentApplicants] = useState([])
   const [interviewSchedule, setInterviewSchedule] = useState([])
   const [rejection, setRejection] = useState([])
   const [blacklisted, setBlacklisted] = useState([])
 
+  const fetchPositions = async () => {
+    try {
+      const response = await api.get(`/company/positions/all`)
+      setPositionsFilter(response.data.positions || [])
+    } catch (error) {
+      console.error("Error fetching job positions:", error)
+      setPositionsFilter([])
+    }
+  }
 
-  const [selectedPosition, setSelectedPosition] = useState(""); // Selected position ID
-
-  // Fetch all dashboard data
   const fetchDashboardData = async () => {
     setLoading(true)
-
     try {
-      // Build query parameters for filtering
       const queryParams = new URLSearchParams()
       if (year) queryParams.append("year", year)
       if (month) queryParams.append("month", month)
       const queryString = queryParams.toString()
 
-      // Replace fetch calls with axios calls
       const [
         summaryResponse,
         statusResponse,
@@ -132,14 +112,13 @@ const AnalysisPage = () => {
         applicantsResponse,
         funnelResponse,
         interviewResponse,
-        //timeResponse,
         rejectionResponse,
       ] = await Promise.all([
         api.get(`/analytics/dashboard/summary?${queryString}`),
         api.get(`/analytics/dashboard/status-distribution?${queryString}`),
         api.get(`/analytics/dashboard/source-distribution?${queryString}`),
         api.get(`/analytics/dashboard/application-source?${queryString}`),
-        api.get(`/analytics/dashboard/monthly-trends?year=${year}`), // Keep this one year-only
+        api.get(`/analytics/dashboard/monthly-trends?year=${year}`),
         api.get(`/analytics/dashboard/job-positions?${queryString}`),
         api.get(`/analytics/dashboard/recent-applicants?${queryString}`),
         api.get(`/analytics/dashboard/hiring-funnel?${queryString}`),
@@ -147,7 +126,6 @@ const AnalysisPage = () => {
         api.get(`/analytic/metrics?${queryString}`),
       ])
 
-      // Update state with fetched data (axios puts response in .data)
       setSummaryData(summaryResponse.data.data)
       setStatusDistribution(statusResponse.data.data)
       setSourceDistribution(sourceResponse.data.data)
@@ -166,99 +144,24 @@ const AnalysisPage = () => {
     }
   }
 
-  const fetchPositions = async () => {
-    try {
-      const response = await api.get(`/company/positions/all`);
-      setPositionsFilter(response.data.positions || []);
-      console.log("Fetched positions:", response.data.positions);
-    } catch (error) {
-      console.error("Error fetching job positions:", error);
-      setPositionsFilter([]); // Reset to empty array on error
-    }
-  };
-  // Fetch job positions data on component mount
-  // Fetch job positions data on component mount
-
-  // Fetch data on component mount
   useEffect(() => {
     const fetchData = async () => {
-      setLoading(true);
+      setLoading(true)
       try {
-        await fetchPositions();
-        await fetchDashboardData();
+        await fetchPositions()
+        await fetchDashboardData()
       } catch (error) {
-        console.error("Error fetching data:", error);
+        console.error("Error fetching data:", error)
       } finally {
-        setLoading(false);
+        setLoading(false)
       }
-    };
+    }
 
-    fetchData();
-  }, [year, month]);
-
-  // Function to handle card expansion
-  const toggleCardExpand = (id) => {
-    const newExpandedState = expandedCard === id ? null : id
-    setExpandedCard(newExpandedState)
-
-    // Toggle body scroll lock when card is expanded
-    document.body.style.overflow = newExpandedState ? "hidden" : "auto"
-  }
+    fetchData()
+  }, [year, month])
 
   const replaceText = (text) => {
     return text ? text.toLowerCase().split('_').map(word => word.charAt(0).toUpperCase() + word.slice(1)).join(' ') : "Null"
-  }
-
-  // Card component with expansion functionality
-  const Card = ({ id, title, icon, children, className = "" }) => {
-    const isExpanded = expandedCard === id
-    const isAnyCardExpanded = expandedCard !== null
-    const shouldHide = isAnyCardExpanded && !isExpanded
-
-    return (
-      <div
-        className={`
-        rounded-xl bg-white/80 backdrop-blur-sm transition-all duration-300 ${className}
-        ${isExpanded
-            ? "fixed top-[10%] left-[10%] right-[10%] bottom-[10%] max-w-5xl mx-auto z-50"
-            : "h-auto min-h-[12rem] xs:min-h-[13rem] sm:min-h-[15rem]"
-          }
-        ${shouldHide ? "opacity-0 pointer-events-none" : "opacity-100"}
-      `}
-        style={{
-          boxShadow: isExpanded ? "0 10px 25px rgba(0, 0, 0, 0.1)" : "0 4px 12px rgba(0, 0, 0, 0.03)",
-          transition: "opacity 0.3s ease, transform 0.3s ease",
-          transform: shouldHide ? "scale(0.95)" : "scale(1)",
-        }}
-      >
-        <div className="flex items-center justify-between p-2 xs:p-3 sm:p-4 border-b border-opacity-10 border-gray-200">
-          <div className="flex items-center gap-1 xs:gap-2 text-xs xs:text-sm sm:text-base font-medium text-gray-700">
-            <span className="text-primary">{icon}</span>
-            {title}
-          </div>
-          <button
-            onClick={() => toggleCardExpand(id)}
-            className="p-1.5 rounded-full hover:bg-gray-100 text-gray-500 transition-colors"
-            aria-label={isExpanded ? "Minimize" : "Expand"}
-          >
-            {isExpanded ? (
-              <FiMinimize2 className="h-3.5 w-3.5 sm:h-4 sm:w-4" />
-            ) : (
-              <FiMaximize2 className="h-3.5 w-3.5 sm:h-4 sm:w-4" />
-            )}
-          </button>
-        </div>
-        <div
-          className={`p-2 xs:p-3 sm:p-4 ${isExpanded
-            ? "h-[calc(100%-60px)] overflow-y-auto"
-            : "h-[calc(100%-40px)] xs:h-[calc(100%-44px)] sm:h-[calc(100%-56px)] custom-scrollbar"
-            }`}
-          style={{ overflowY: "auto" }}
-        >
-          {children}
-        </div>
-      </div>
-    )
   }
 
   // Prepare data for monthly trends chart
@@ -269,63 +172,19 @@ const AnalysisPage = () => {
         label: "Applicants",
         data: monthlyTrends.map((item) => item.applicant_count),
         fill: true,
-        backgroundColor: "rgba(0, 128, 128, 0.2)", // teal with opacity
-        borderColor: "rgba(0, 128, 128, 1)", // teal
+        backgroundColor: "rgba(0, 128, 128, 0.2)",
+        borderColor: "rgba(0, 128, 128, 1)",
         tension: 0.4,
       },
       {
         label: "Hired",
         data: monthlyTrends.map((item) => item.hired_count),
         fill: true,
-        backgroundColor: "rgba(0, 96, 96, 0.2)", // darker teal with opacity
-        borderColor: "rgba(0, 96, 96, 1)", // darker teal
+        backgroundColor: "rgba(0, 96, 96, 0.2)",
+        borderColor: "rgba(0, 96, 96, 1)",
         tension: 0.4,
       },
     ],
-  }
-
-  // Wrapper for chart cards to make them expandable
-  const ChartCard = ({ id, title, subtitle, icon, children }) => {
-    const isExpanded = expandedCard === id
-    const isAnyCardExpanded = expandedCard !== null
-    const shouldHide = isAnyCardExpanded && !isExpanded
-
-    return (
-      <div
-        className={`
-          bg-white rounded-xl border border-gray-200 shadow-sm overflow-hidden transition-all duration-300
-          ${isExpanded ? "fixed top-[10%] left-[10%] right-[10%] bottom-[10%] max-w-5xl mx-auto z-50" : "h-auto"}
-          ${shouldHide ? "opacity-0 pointer-events-none" : "opacity-100"}
-        `}
-        style={{
-          boxShadow: isExpanded ? "0 10px 25px rgba(0, 0, 0, 0.1)" : "0 4px 12px rgba(0, 0, 0, 0.03)",
-          transition: "opacity 0.3s ease, transform 0.3s ease",
-          transform: shouldHide ? "scale(0.95)" : "scale(1)",
-        }}
-      >
-        <div className="flex items-center justify-between px-4 py-3 sm:px-6 sm:py-4 border-b border-gray-200">
-          <div>
-            <div className="flex items-center gap-2 font-semibold text-gray-900">
-              <span className="text-primary">{icon}</span>
-              <h3 className="text-base sm:text-lg">{title}</h3>
-            </div>
-            {subtitle && <p className="text-xs sm:text-sm text-gray-500 mt-1">{subtitle}</p>}
-          </div>
-          <button
-            onClick={() => toggleCardExpand(id)}
-            className="p-1.5 rounded-full hover:bg-gray-100 text-gray-500 transition-colors"
-            aria-label={isExpanded ? "Minimize" : "Expand"}
-          >
-            {isExpanded ? (
-              <FiMinimize2 className="h-3.5 w-3.5 sm:h-4 sm:w-4" />
-            ) : (
-              <FiMaximize2 className="h-3.5 w-3.5 sm:h-4 sm:w-4" />
-            )}
-          </button>
-        </div>
-        <div className={`p-4 sm:p-6 ${isExpanded ? "h-[calc(100%-70px)] overflow-y-auto" : ""}`}>{children}</div>
-      </div>
-    )
   }
 
   return (
@@ -357,115 +216,73 @@ const AnalysisPage = () => {
           scrollbar-color: rgba(0, 0, 0, 0.1) transparent;
         }
       `}</style>
-      {/* Filter controls */}
-      <div className="flex flex-wrap items-center gap-3 mb-4">
-        <div className="flex items-center gap-2">
-          <FiCalendar className="text-primary h-4 w-4" />
-          <span className="text-sm font-medium">Filters:</span>
-        </div>
 
-        <div className="flex flex-wrap gap-2">
-          {/* Year Select */}
-          <select
-            value={year}
-            onChange={(e) => setYear(e.target.value)}
-            className="px-3 py-1.5 text-sm border border-gray-300 rounded-md bg-white shadow-sm focus:outline-none focus:ring-1 focus:ring-primary"
-          >
-            {yearOptions.map((option) => (
-              <option key={option.value} value={option.value}>
-                {option.label}
-              </option>
-            ))}
-          </select>
+      <FilterControls
+        year={year}
+        setYear={setYear}
+        month={month}
+        setMonth={setMonth}
+        selectedPosition={selectedPosition}
+        setSelectedPosition={setSelectedPosition}
+        positionsFilter={positionsFilter}
+        fetchDashboardData={fetchDashboardData}
+      />
 
-          {/* Month Select */}
-          <select
-            value={month}
-            onChange={(e) => setMonth(e.target.value)}
-            className="px-3 py-1.5 text-sm border border-gray-300 rounded-md bg-white shadow-sm focus:outline-none focus:ring-1 focus:ring-primary"
-          >
-            {monthOptions.map((option) => (
-              <option key={option.value} value={option.value}>
-                {option.label}
-              </option>
-            ))}
-          </select>
-
-          {/* Position Select */}
-          <select
-            value={selectedPosition} // Use the new state for the selected position
-            onChange={(e) => setSelectedPosition(e.target.value)} // Update the selected position
-            className="px-3 py-1.5 text-sm border border-gray-300 rounded-md bg-white shadow-sm focus:outline-none focus:ring-1 focus:ring-primary"
-            disabled={!positionsFilter || positionsFilter.length === 0} // Disable if no positions
-          >
-            <option value="">All Positions</option>
-            {positionsFilter && positionsFilter.length > 0 ? (
-              positionsFilter.map((position) => (
-                <option key={position._id} value={position._id}>
-                  {position.title}
-                </option>
-              ))
-            ) : (
-              <option value="" disabled>Loading positions...</option>
-            )}
-          </select>
-          {/* Reset Button */}
-          <button
-            onClick={() => {
-              setYear(new Date().getFullYear().toString());
-              setMonth("");
-              setSelectedPosition(""); // Reset the selected position
-              fetchDashboardData();
-            }}
-            className="px-3 py-1.5 text-sm bg-gray-100 hover:bg-gray-200 rounded-md flex items-center gap-1 transition-colors"
-          >
-            <FiRefreshCw className="h-3.5 w-3.5" />
-            Reset
-          </button>
-        </div>
-      </div>
-
-      {/* Overlay for expanded card */}
-      {expandedCard && (
-        <div
-          className="fixed inset-0 bg-black/50 z-50 flex items-center justify-center min-h-screen"
-          onClick={() => {
-            setExpandedCard(null)
-            document.body.style.overflow = "auto"
-          }}
-        />
-      )}
+      <Overlay expandedCard={expandedCard} setExpandedCard={setExpandedCard} />
 
       {/* Top row with 4 equal cards */}
-      <div
-        className={`grid grid-cols-1 gap-3 xs:gap-4 sm:gap-5 md:gap-6 sm:grid-cols-2 lg:grid-cols-4 ${expandedCard ? "z-0" : ""}`}
-      >
-        <Card id="applications" title="Applications Received" icon={<FiUsers className="h-4 w-4 sm:h-5 sm:w-5" />}>
+      <div className={`grid grid-cols-1 gap-3 xs:gap-4 sm:gap-5 md:gap-6 sm:grid-cols-2 lg:grid-cols-4 ${expandedCard ? "z-0" : ""}`}>
+        <Card 
+          id="applications" 
+          title="Applications Received" 
+          icon={<FiUsers className="h-4 w-4 sm:h-5 sm:w-5" />}
+          expandedCard={expandedCard}
+          setExpandedCard={setExpandedCard}
+        >
           <ApplicationReceived year={year} month={month} isExpanded={expandedCard === "applications"} />
         </Card>
 
-        <Card id="positions" title="Top Job Positions" icon={<FiBriefcase className="h-4 w-4 sm:h-5 sm:w-5" />}>
+        <Card 
+          id="positions" 
+          title="Top Job Positions" 
+          icon={<FiBriefcase className="h-4 w-4 sm:h-5 sm:w-5" />}
+          expandedCard={expandedCard}
+          setExpandedCard={setExpandedCard}
+        >
           <TopJobPositions jobPositions={jobPositions} loading={loading} year={year} month={month} isExpanded={expandedCard === "positions"} />
         </Card>
 
-        <Card id="hires" title="Internal vs External Hires" icon={<FiRefreshCw className="h-4 w-4 sm:h-5 sm:w-5" />}>
+        <Card 
+          id="hires" 
+          title="Internal vs External Hires" 
+          icon={<FiRefreshCw className="h-4 w-4 sm:h-5 sm:w-5" />}
+          expandedCard={expandedCard}
+          setExpandedCard={setExpandedCard}
+        >
           <InternalVsExternalHires year={year} month={month} />
         </Card>
 
-        <Card id="dropoff" title="Candidate Drop-off Rate" icon={<FiTrendingDown className="h-4 w-4 sm:h-5 sm:w-5" />}>
+        <Card 
+          id="dropoff" 
+          title="Candidate Drop-off Rate" 
+          icon={<FiTrendingDown className="h-4 w-4 sm:h-5 sm:w-5" />}
+          expandedCard={expandedCard}
+          setExpandedCard={setExpandedCard}
+        >
           <CandidateDropOffRate year={year} month={month} isExpanded={expandedCard === "dropoff"} />
         </Card>
       </div>
 
       {/* Charts Section */}
       <div className="grid grid-cols-1 lg:grid-cols-2 gap-6 mb-6">
-
         {/* Job Position Analytics */}
         <ChartCard
           id="positionAnalytics"
           title="Top Job Positions"
           subtitle="Applicants by job position"
           icon={<FiBriefcase className="h-4 w-4 sm:h-5 sm:w-5" />}
+          expandedCard={expandedCard}
+          setExpandedCard={setExpandedCard}
         >
           {loading ? (
             <div className="w-full h-[300px] flex items-center justify-center">
@@ -480,12 +297,12 @@ const AnalysisPage = () => {
                     {
                       label: "Applicants",
                       data: jobPositions.map((item) => item.applicant_count),
-                      backgroundColor: "rgba(0, 128, 128, 0.8)", // teal
+                      backgroundColor: "rgba(0, 128, 128, 0.8)",
                     },
                     {
                       label: "Hired",
                       data: jobPositions.map((item) => item.hired_count),
-                      backgroundColor: "rgba(0, 96, 96, 0.8)", // darker teal
+                      backgroundColor: "rgba(0, 96, 96, 0.8)",
                     },
                   ],
                 }}
@@ -508,8 +325,8 @@ const AnalysisPage = () => {
                           size: 11,
                         },
                         callback: function (value) {
-                          const label = this.getLabelForValue(value);
-                          return label.split(' ');
+                          const label = this.getLabelForValue(value)
+                          return label.split(' ')
                         },
                       },
                     },
@@ -529,6 +346,8 @@ const AnalysisPage = () => {
           title="Hiring Funnel"
           subtitle="Applicants at each stage of the hiring process"
           icon={<FiUsers className="h-4 w-4 sm:h-5 sm:w-5" />}
+          expandedCard={expandedCard}
+          setExpandedCard={setExpandedCard}
         >
           {loading ? (
             <div className="w-full h-[300px] flex items-center justify-center">
@@ -536,7 +355,6 @@ const AnalysisPage = () => {
             </div>
           ) : (
             <div className="h-[600px]">
-              {/* <TopJobPositions jobPositions={jobPositions} loading={loading} year={year} month={month} /> */}
               <HiringFunnelChart data={hiringFunnel} />
             </div>
           )}
@@ -551,6 +369,8 @@ const AnalysisPage = () => {
             title="Requisition Analysis"
             subtitle="Requisition Data"
             icon={<FiBarChart2 className="h-4 w-4 sm:h-5 sm:w-5" />}
+            expandedCard={expandedCard}
+            setExpandedCard={setExpandedCard}
           >
             <div className="w-full h-[400px]">
               <ApplicantStatusChart year={year} month={month} className />
@@ -564,6 +384,8 @@ const AnalysisPage = () => {
             title="Monthly Applicant Trends"
             subtitle="Applications and hires by month"
             icon={<FiBarChart2 className="h-4 w-4 sm:h-5 sm:w-5" />}
+            expandedCard={expandedCard}
+            setExpandedCard={setExpandedCard}
           >
             {loading ? (
               <div className="w-full h-[300px] flex items-center justify-center">
@@ -606,6 +428,8 @@ const AnalysisPage = () => {
           title="Applicant Sources"
           subtitle="Where applicants applied from"
           icon={<FiPieChart className="h-4 w-4 sm:h-5 sm:w-5" />}
+          expandedCard={expandedCard}
+          setExpandedCard={setExpandedCard}
         >
           {loading ? (
             <div className="w-full h-[300px] flex items-center justify-center">
@@ -631,7 +455,7 @@ const AnalysisPage = () => {
                     maintainAspectRatio: false,
                     plugins: {
                       legend: {
-                        display: true, // Hides the legend labels
+                        display: true,
                       },
                       tooltip: {
                         callbacks: {
@@ -669,6 +493,8 @@ const AnalysisPage = () => {
           title="Source of Application"
           subtitle="Where applicants discovered this"
           icon={<FiPieChart className="h-4 w-4 sm:h-5 sm:w-5" />}
+          expandedCard={expandedCard}
+          setExpandedCard={setExpandedCard}
         >
           {loading ? (
             <div className="w-full h-[300px] flex items-center justify-center">
@@ -694,7 +520,7 @@ const AnalysisPage = () => {
                     maintainAspectRatio: false,
                     plugins: {
                       legend: {
-                        display: true, // Hides the legend labels
+                        display: true,
                       },
                       tooltip: {
                         callbacks: {
@@ -732,6 +558,8 @@ const AnalysisPage = () => {
           title="Rejected"
           subtitle="Reason for Rejection"
           icon={<FiPieChart className="h-4 w-4 sm:h-5 sm:w-5" />}
+          expandedCard={expandedCard}
+          setExpandedCard={setExpandedCard}
         >
           {loading ? (
             <div className="w-full h-[300px] flex items-center justify-center">
@@ -757,7 +585,7 @@ const AnalysisPage = () => {
                     maintainAspectRatio: false,
                     plugins: {
                       legend: {
-                        display: true, // Hides the legend labels
+                        display: true,
                       },
                       tooltip: {
                         callbacks: {
@@ -795,6 +623,8 @@ const AnalysisPage = () => {
           title="Blacklisted"
           subtitle="Reason for Blacklisted"
           icon={<FiPieChart className="h-4 w-4 sm:h-5 sm:w-5" />}
+          expandedCard={expandedCard}
+          setExpandedCard={setExpandedCard}
         >
           {loading ? (
             <div className="w-full h-[300px] flex items-center justify-center">
@@ -820,7 +650,7 @@ const AnalysisPage = () => {
                     maintainAspectRatio: false,
                     plugins: {
                       legend: {
-                        display: true, // Hides the legend labels
+                        display: true,
                       },
                       tooltip: {
                         callbacks: {
@@ -852,18 +682,6 @@ const AnalysisPage = () => {
             </div>
           )}
         </ChartCard>
-
-
-        {/* <ChartCard 
-        id="source" 
-        title="Source of Applications" 
-        icon={<FiPieChart className="h-4 w-4 sm:h-5 sm:w-5" />}
-        subtitle="Where applicants applied"
-        >
-            <div className="w-full h-[300px]">
-              <SourceOfApplication year={year} month={month} />
-            </div>    
-        </ChartCard> */}
       </div>
     </div>
   )
